extract into legacy files
